+++ src/js/browserify/path/index.mjs
@@ -1,0 +1,5 @@
+import path from "path-browserify"; // Latest version which supports `path.parse`
+
+const {dirname, join, parse, sep} = path;
+
+export {dirname, join, parse, sep};