+++ src/js/worker.js
@@ -1,17 +1,17 @@
 import {
     AbstractLog,
     ArrayInput,
-    ConvertJavaTextureToBedrock,
+    ConvertJavaTextureToBedrockApi,
     FileInputEntry,
     FileOutput
-} from "@ozelot379/convert-minecraft-java-texture-to-bedrock";
+} from "@ozelot379/convert-minecraft-java-texture-to-bedrock-api";
 
 addEventListener("message", async (e) => {
     try {
         const {files, options} = e.data;
 
         const output = {
-            output: await new ConvertJavaTextureToBedrock(
+            output: await new ConvertJavaTextureToBedrockApi(
                 new ArrayInput(Array.prototype.map.call(files, file => new FileInputEntry(file))),
                 new FileOutput(),
                 new class extends AbstractLog {