+++ CHANGELOG.md
@@ -1,3 +1,6 @@
 # Changelog
 
-Look at https://github.com/ozelot379/ConvertJavaTextureToBedrockApi/blob/master/CHANGELOG.md
+## [3.5.0]
+- Convert dolphin texture
+
+Changelog for previous versions are available at https://github.com/ozelot379/ConvertJavaTextureToBedrockApi/blob/master/CHANGELOG.md#341